src/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
python-multipart==0.0.20
pydantic-settings==2.12.0
aiofiles==25.1.0
langchain==1.1.0
langchain-community==0.4.1
PyMuPDF==1.26.6
docx2txt==0.9
python-pptx==1.0.2
unstructured==0.18.21
